chore: ForeignEntityTest 추가, EntityTest 구현 중

+++ exposed/domain/src/test/kotlin/io/bluetape4k/workshop/exposed/domain/WithSchemas.kt
@@ -22,12 +22,15 @@ fun withSchemas(
 
     settings.forEach { dialect ->
         withDb(dialect, configure) {
-            SchemaUtils.createSchema(*schemas)
-            try {
-                statement()
-            } finally {
-                SchemaUtils.dropSchema(*schemas, cascade = true)
-                commit()
+            if (currentDialectTest.supportsCreateSchema) {
+                SchemaUtils.createSchema(*schemas)
+                try {
+                    statement()
+                    commit()     // Need commit to persist data before drop schemas
+                } finally {
+                    SchemaUtils.dropSchema(*schemas, cascade = true)
+                    commit()
+                }
             }
         }
     }
